requirements.txt
bashlex
pytest
pyyaml
black
flake8